fix: ts problems
