+++ src/layout/dagre/src/graphlib.ts
@@ -1,4 +1,4 @@
-
+// @ts-ignore
 import glib from '@dagrejs/graphlib';
 
 let graphlib = glib;
