+++ src/layout/dagre/src/rank/feasible-tree.ts
@@ -1,14 +1,9 @@
-// "use strict";
-
-// const _ = require("../lodash");
-// const Graph = require("../graphlib").Graph;
-// const slack = require("./util").slack;
-
-import { Graph as IGraph } from '../../types';
+import { graphlib as IGraphLib } from '../../graphlib';
 import graphlib from '../graphlib';
 import { slack } from './util';
 import { minBy } from '../util';
 
+type IGraph = IGraphLib.Graph;
 const Graph = (graphlib as any).Graph;
